[MOD] modificado el helper drop_cascade para que pueda eliminar dinamicamente sin dar error cuando no exista la tabla

+++ app/helpers/helpers_db.php
@@ -2,5 +2,8 @@
 
 
 function drop_cascade($table = null){
-	DB::statement('DROP TABLE '.$table.' CASCADE;');
+	if( Schema::hasTable($table) ){
+		DB::statement('DROP TABLE '.$table.' CASCADE;');
+	}
+
 }
